pyproject.toml
[project]
name = "py-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.4",
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ty>=0.0.14",
]
